Add conversion for numeric comparisons
