+++ business_rules/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.3.2"
+__version__ = "1.3.3"
 
 from .engine import run_all
 from .utils import export_rule_data
